requirements.txt
asgiref==3.8.1
Django==3.2
django-cors-headers==4.4.0
djangorestframework==3.12.4
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.1
typing_extensions==4.12.2
